+++ app/src/main/java/com/techcourse/ManualHandlerAdapter.java
@@ -1,6 +1,5 @@
 package com.techcourse;
 
-import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
